Added interface functionality to Car and subclasses. Added inheritance and correct JSON syntax to make the distinction between the classes.
Switched BigDecimal to double for now...

+++ src/main/java/com/example/sharemobility/Calculator.java
@@ -1,6 +1,8 @@
 package com.example.sharemobility;
 
+import java.math.BigDecimal;
+
 public interface Calculator {
-    float calculateTCO();
-    float calculateUsageCostsPerKilometer();
+    double calculateTCO();
+    double calculateUsageCostsPerKilometer();
 }
